+++ core/engine.h
@@ -54,10 +54,10 @@ OIDN_NAMESPACE_BEGIN
 
     virtual Ref<Buffer> newNativeBuffer(void* handle);
 
-    virtual Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlag fdType,
+    virtual Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlags fdType,
                                           int fd, size_t byteSize);
 
-    virtual Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlag handleType,
+    virtual Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlags handleType,
                                           void* handle, const void* name, size_t byteSize);
     // Tensor
     virtual bool isSupported(const TensorDesc& desc) const;
@@ -65,10 +65,10 @@ OIDN_NAMESPACE_BEGIN
     virtual Ref<Tensor> newTensor(const Ref<Buffer>& buffer, const TensorDesc& desc, size_t byteOffset = 0);
 
     // Semaphore
-    virtual Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlag fdType,
+    virtual Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlags fdType,
                                                 int fd);
 
-    virtual Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlag handleType,
+    virtual Ref<Semaphore> newExternalSemaphore(ExternalSemaphoreTypeFlags handleType,
                                                 void* handle, const void* name);
 
     virtual void submitSignalSemaphores(Semaphore* const* semaphores,