+++ devices/hip/hip_engine.h
@@ -64,10 +64,10 @@ OIDN_NAMESPACE_BEGIN
     hipStream_t getHIPStream() const { return stream; }
 
     // Buffer
-    Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlag fdType,
+    Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlags fdType,
                                   int fd, size_t byteSize) override;
 
-    Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlag handleType,
+    Ref<Buffer> newExternalBuffer(ExternalMemoryTypeFlags handleType,
                                   void* handle, const void* name, size_t byteSize) override;
 
     // Tensor