+++ devices/sycl/sycl_device.cpp
@@ -439,6 +439,8 @@ OIDN_NAMESPACE_BEGIN
     #else
       externalMemoryTypes = ExternalMemoryTypeFlag::DMABuf;
     #endif
+
+      externalMemoryTypes |= ExternalMemoryTypeFlag::Dedicated;
     }
   }